Minor: Added fixture to test_pyfdb
Now the data is written to a temp fdb home directory instead of the data
root of the repository

Minor: Changes for CICD
- Fixes for Isort
- Changes for black
- Fixes flake8
- Added requirements.txt
  The CICD is relying on the requirements.txt.

Minor: Added test for grib comp/tmp dir generator

FDB space is not setup for each test individually in a temp directory.

Minor: Extended test cases for checking of written files

Individually written grib messages are cross-checked with the output of
a FDB request which retrieves the two fields combined.

Minor: Restructuring for the CICD
- Fixes for python-qa
- Work around for python 3.10 in CICD
  The delete parameter of the TemporaryDirectory function was introduced
  in python 3.12. So now the creation and deletion is handled by the
  corresponding pytest function
- Added grib files back to repository
  Those were ignored after moving due to a gitignore entry.
  Adjusted the entry.
- Switched back to gitroot to determine data files