Minor: Added LD_LIBRARY_PATH and DYLD_LIBRARY_PATH to tox
Those env vars are now set in tox and therefore tox works without
further issues and is able to find the locally installed FDB.

Eccodes since 2.37.0 has the option to run without the shipped binaries.
Therefore we disable the install of bundled eccodes.so and use the one
needed for the compilation of the whole stack.